Refactor des noms et types des fonctions
