+++ README.md
@@ -2,6 +2,10 @@
 
 This script aims to automatically upscale comics it finds in a folder. It uses Real-ESRGAN to upscale the images.
 
+The comics can be in a folder, CBZ, or any ZIP file format.
+Most image formats are supported, including PNG, JPEG, and WEBP.
+Animated WEBP are not supported and will be converted to a static image (the first frame).
+
 ## Requirements
 
 Use conda to install the requirements:
